% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_map_col.R
\name{add_map_col}
\alias{add_map_col}
\title{Adds a new column to a tidy data.frame from a map with a common column.}
\usage{
add_map_col(data, map_tib, common_col)
}
\arguments{
\item{data}{A data.frame with at least column \code{common_col}}

\item{map_tib}{A data.frame with two columns, one being \code{common_col},
and the other being the new column to add.}

\item{common_col}{A string. The common column as basis of the mapping.}
}
\value{
The data.frame with the new mapped column.
}
\description{
Adds a new column to a tidy data.frame from a map with a common column.
}
