% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_num_id_col.R
\name{add_num_id_col}
\alias{add_num_id_col}
\title{Adds a column with numeric IDs to a tidy data.frame}
\usage{
add_num_id_col(data)
}
\arguments{
\item{data}{A data.frame with at least column \code{ID}}
}
\value{
data A data.frame with at least column \code{ID}, and \code{ID_num}
}
\description{
Adds a column with numeric IDs to a tidy data.frame
}
\examples{
data <- create_test_df_multi_icd_ver()
data_with_num_id <- add_num_id_col(data)

}
