% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cci.R
\name{calc_cci}
\alias{calc_cci}
\title{Calculates the Charlson Comorbidity Index, based on longitudinal data
in INTERVENE format.}
\usage{
calc_cci(data)
}
\arguments{
\item{data}{A data.frame with at least column \code{ID}, \code{EVENT_AGE},
and \code{PRIMARY_ICD}.}
}
\value{
A tibble with columns \code{ID} and \code{score} with the charlson
weighted comorbidity scores for each individual.
}
\description{
Maps data onto the relevant conditions for the charlson-deyo
comorbidity index, using \code{\link[comorbidity]{comorbidity}}.
Then calculates the scores, using \code{\link[comorbidity]{score}}
}
\details{
Can handle different ICD-versions also for a single patient, by adding
up the results.
}
