% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_df_multi_icd_ver.R
\name{create_test_df_multi_icd_ver}
\alias{create_test_df_multi_icd_ver}
\title{Creates a tibble with random ICD-codes entries, from different ICD versions,
using \code{\link{create_test_df}}. Meant for testing.}
\usage{
create_test_df_multi_icd_ver(
  n_icd10 = 50,
  n_icd9 = 20,
  icd10_indv = c("FG0000001", "FG0000002", "FG0000004", "FG0000005", "FG0000006"),
  icd9_indv = c("FG0000005", "FG0000007", "FG0000002", "FG0000008", "FG0000009")
)
}
\arguments{
\item{n_icd10}{Number of draws from ICD10}

\item{n_icd9}{Number of draws from ICD9}

\item{icd10_indv}{A character vector.
The names of the individuals with ICD10 codes}

\item{icd9_indv}{A character vector.
The names of the individuals with ICD9 codes}
}
\value{
A tibble with columns \code{ID}, \code{Event_age}, \code{primary_ICD},
and \code{secondary_ICD} and random ICD codes, as well as age at event.
}
\description{
It is possible to create a mix of ICD-10 and ICD-9 codes to create
a more realistic testing dataset.
}
