% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_icd_ver.R
\name{group_by_icd_ver}
\alias{group_by_icd_ver}
\title{Groups the data according to the ICD-versions.}
\usage{
group_by_icd_ver(data)
}
\arguments{
\item{data}{A tidy data.frame with at least column \code{ICD_VERSION}.}
}
\value{
A list c(\code{groups}, \code{group_keys}, \code{group_idxs}):
- \code{groups} The grouped tibble.
- \code{group_keys} A list.
The name of each group, so the ICD-versions
present in the data.
- \code{group_idxs} A named list.
The indices of the elements of each group
in the original data. The names are the group names.
}
\description{
Groups the data according to the ICD-versions.
}
