% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_id_age.R
\name{sort_by_id_age}
\alias{sort_by_id_age}
\title{Sorts a tidy data.frame by \code{ID} and \code{EVENT_AGE}.}
\usage{
sort_by_id_age(data)
}
\arguments{
\item{data}{A data.frame with at least column \code{ID}, and \code{EVENT_AGE}}
}
\value{
The data.drame sorted first by \code{ID}, and then by \code{EVENT_AGE}.
}
\description{
Sorts a tidy data.frame by \code{ID} and \code{EVENT_AGE}.
}
